requirements.txt
streamlit>=1.41.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.0.0
ddgs>=9.0.0
groq>=0.4.0
google-generativeai>=0.8.0
huggingface_hub>=0.20.0
nltk>=3.8.0
scikit-learn>=1.3